requirements.txt
numpy
pandas
matplotlib
streamlit
requests
folium
python-dotenv
streamlit-folium



